chore(nuz-cli): rename functions in user commands
