+++ packages/nuz-cli/src/commands/user/logoutFromUser.ts
